The type of the left variable does not match the type of the right variable.
Refer the context around the expression to decide either to change the
type of the left variable or use "static_cast<>()" to get the value of
compatible type from the right varible.

+++ native_engine/impl/ark/native_value/ark_native_string.cpp
@@ -78,7 +78,7 @@ size_t ArkNativeString::EncodeWriteUtf8(char* buffer, size_t bufferSize, int32_t
     int32_t length = val->Length();
 
     int32_t pos = 0;
-    int32_t writableSize = bufferSize;
+    int32_t writableSize = static_cast<int32_t>(bufferSize);
     int32_t i = 0;
     Local<ObjectRef> strObj = Local<ObjectRef>(val.ToLocal(vm));
     for (; i < length; i++) {
